refactor: centralize integer parsing
