+++ src/gateway/gateway-http-utils.ts
@@ -7,19 +7,11 @@
 import type { IncomingMessage, ServerResponse } from 'node:http';
 
 import { GatewayRequestError } from '../errors/gateway-request-error.js';
+import { parsePositiveInteger } from '../utils/number-normalization.js';
 
 const MAX_REQUEST_BYTES = 1_000_000; // 1 MB
 
-export function parsePositiveInteger(value: unknown): number | null {
-  if (typeof value === 'number') {
-    return Number.isInteger(value) && value > 0 ? value : null;
-  }
-  if (typeof value !== 'string') return null;
-  const normalized = value.trim();
-  if (!/^\d+$/.test(normalized)) return null;
-  const parsed = Number(normalized);
-  return Number.isSafeInteger(parsed) && parsed > 0 ? parsed : null;
-}
+export { parsePositiveInteger };
 
 export async function readRequestBody(
   req: IncomingMessage,
